+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Changelog
 * Rename ``--samples`` argument from :command:`compute-copy-number` command to ``--samples-without-sv``.
 * Rename ``--samples`` argument from :command:`run-ngs-pipeline` command to ``--samples-without-sv``.
 * Update :command:`run-ngs-pipeline` and :command:`run-chip-pipeline` commands to be able to subset/exclude specified samples.
+* Remove ``--fn`` argument from :command:`filter-samples` command.
 
 0.9.0 (2021-12-05)
 ------------------